Fix title emphasis in README
Markdown requires a space between the hash symbol (`#`) and the header text.
This was not the case for two of the headers in the README. I have added the
missing spaces.

Furthermore, markdown supports multiple header levels. I've reduced the level
for the subtitles while keeping the title the same.